Initialize early some message types
