+++ apps/meteor/app/livechat/client/index.ts
@@ -1,3 +1,1 @@
-import '../lib/messageTypes';
 import './voip';
-import './ui';
