+++ ee/packages/omnichannel-services/package.json
@@ -17,6 +17,7 @@
 		"@rocket.chat/emitter": "~0.31.25",
 		"@rocket.chat/i18n": "workspace:^",
 		"@rocket.chat/logger": "workspace:^",
+		"@rocket.chat/message-types": "workspace:~",
 		"@rocket.chat/model-typings": "workspace:^",
 		"@rocket.chat/models": "workspace:^",
 		"@rocket.chat/pdf-worker": "workspace:^",
