+++ packages/message-types/package.json
@@ -16,12 +16,18 @@
 	"devDependencies": {
 		"@rocket.chat/core-typings": "workspace:~",
 		"@rocket.chat/eslint-config": "workspace:~",
+		"date-fns": "~4.1.0",
 		"eslint": "~8.45.0",
 		"i18next": "~23.4.9",
 		"jest": "~30.0.5",
+		"moment": "^2.30.1",
 		"typescript": "~5.9.2"
 	},
 	"volta": {
 		"extends": "../../package.json"
+	},
+	"peerDependencies": {
+		"date-fns": "~4.1.0",
+		"moment": "^2.30.1"
 	}
 }
